Fix portability issue with printing
The gene print function of PGApack gets a FILE * as parameter.
Previously we used dup2 to duplicate the underlying file descriptor in
Python3. Since this is non-portable we now keep copies of the FILE * and
the generated python file object inside the PGA object. This avoids
both, garbage collection and memory leak issues and is more portable.