Ensure destructor is called
We had a dictionary with all the PGA objects. This prevented the
destructor -- and the PGA_Destroy which in turn would call the
MPI_Finalize -- from being called. This made problems with parallel
versions where we would get an error message and sometimes lose part of
the output (the description of the best string, the purpose of the whole
optimization!). We now use a WeakValueDictionary and use weak references
for the PGA objects in the contexts dictionary.

Also add a print_context debug method.