Hillclimbing and deterministic random generator
Upstream added a user function for hillclimbing, we're implementing this
with a new 'hillclimb' member function. See examples/one_max_hc.py
When using random numbers in the hillclimber we want to retain the
feature that runs are reproduceable when a random seed is specified.
This is done with the new constructor parameter 'random_deterministic'.

All examples computed the best individual but re-computed the
evaluation. We now call get_evaluation instead.